+++ backend/application/api/src/main/kotlin/io/raemian/api/goal/model/GoalResult.kt
@@ -1,6 +1,6 @@
 package io.raemian.api.goal.model
 
-import io.raemian.api.support.utils.format
+import io.raemian.api.support.extension.format
 import io.raemian.storage.db.core.goal.Goal
 import io.raemian.storage.db.core.tag.Tag
 import io.raemian.storage.db.core.task.Task
