+++ backend/application/api/src/main/kotlin/io/raemian/api/support/extension/LocalDateExtensionFunctions.kt
@@ -1,0 +1,12 @@
+package io.raemian.api.support.extension
+
+import java.time.LocalDate
+
+fun LocalDate.format(): String {
+    var month = (this.monthValue).toString()
+    if (month.length == 1) {
+        month = "0$month"
+    }
+
+    return "${this.year}.$month"
+}
