+++ backend/application/api/src/test/kotlin/io/raemian/api/integration/goal/GoalServiceTest.kt
@@ -5,7 +5,7 @@ import io.raemian.api.goal.controller.request.UpdateGoalRequest
 import io.raemian.api.goal.service.GoalService
 import io.raemian.api.support.exception.MaxGoalCountExceededException
 import io.raemian.api.support.exception.PrivateLifeMapException
-import io.raemian.api.support.utils.RaemianLocalDateUtil
+import io.raemian.api.support.utils.DeadlineCreator
 import io.raemian.storage.db.core.goal.Goal
 import io.raemian.storage.db.core.goal.GoalRepository
 import io.raemian.storage.db.core.lifemap.LifeMap
@@ -243,7 +243,7 @@ class GoalServiceTest {
         val description = "Description"
         val year = "2000"
         val month = "05"
-        val deadline = RaemianLocalDateUtil.of(year, month)
+        val deadline = DeadlineCreator.create(year, month)
 
         val goal = Goal(
             lifeMap = LIFE_MAP_FIXTURE,
